application: serial_lte_modem: new nRF Cloud service
Separate nRF Cloud service from the GNSS service to allow
include/exclude of nRF Cloud services by configurating prj.conf,
and allow more nRF Cloud-specific services be added later.

New configuration CONFIG_SLM_NRF_CLOUD for nRF Cloud access.
nRF Cloud A-GPS, P-GPS and Location service to be configured
separately with lib_nrfcloud configuration items:
.Remove CONFIG_SLM_LOCATION, to use CONFIG_NRF_CLOUD_LOCATION.
.Remove CONFIG_SLM_AGPS, to use CONFIG_NRF_CLOUD_AGPS.
.Remove CONFIG_SLM_PGPS, to use CONFIG_NRF_CLOUD_PGPS.

Enable nRF Cloud A-GPS and Location service by default.
