+++ applications/serial_lte_modem/doc/AT_commands_intro.rst
@@ -47,3 +47,4 @@ The modem-specific AT commands are documented in the `nRF91 AT Commands Referenc
    TWI_AT_commands
    GPIO_AT_commands
    CARRIER_AT_commands
+   NRFCLOUD_AT_commands
