+++ applications/serial_lte_modem/overlay-carrier.conf
@@ -34,7 +34,3 @@ CONFIG_PM_PARTITION_SIZE_NVS_STORAGE=0
 
 # Enable LwM2M carrier support in SLM
 CONFIG_SLM_CARRIER=y
-
-# Disable PGPS to have enough space for carrier library
-CONFIG_SLM_PGPS=n
-CONFIG_NRF_CLOUD_PGPS=n
