+++ applications/serial_lte_modem/prj.conf
@@ -74,21 +74,22 @@ CONFIG_SETTINGS_FCB=y
 CONFIG_FCB=y
 
 # nRF Cloud
-#CONFIG_NRF_CLOUD_IPV6=y
+CONFIG_NRF_CLOUD_IPV6=n
 CONFIG_NRF_CLOUD_MQTT=y
 CONFIG_NRF_CLOUD_FOTA=n
 CONFIG_NRF_CLOUD_AGPS=y
 CONFIG_NRF_CLOUD_AGPS_FILTERED=n
-CONFIG_NRF_CLOUD_PGPS=y
+CONFIG_NRF_CLOUD_PGPS=n
 CONFIG_NRF_CLOUD_LOCATION=y
 CONFIG_NRF_CLOUD_LOG_LEVEL_INF=y
 CONFIG_NRF_CLOUD_GPS_LOG_LEVEL_INF=y
+CONFIG_NRF_CLOUD_CLIENT_ID_SRC_INTERNAL_UUID=y
 CONFIG_DATE_TIME=y
 CONFIG_MODEM_INFO=y
 CONFIG_MODEM_INFO_ADD_DATE_TIME=n
-CONFIG_AT_MONITOR=y
+
+# UUID and JWT
 CONFIG_MODEM_JWT=y
-CONFIG_NRF_CLOUD_CLIENT_ID_SRC_INTERNAL_UUID=y
 
 #
 # SLM-specific configurations
@@ -98,8 +99,3 @@ CONFIG_SLM_LOG_LEVEL_INF=y
 CONFIG_SLM_EXTERNAL_XTAL=n
 CONFIG_SLM_START_SLEEP=n
 CONFIG_SLM_DATAMODE_URC=n
-
-# nRF Cloud based location services
-CONFIG_SLM_AGPS=n
-CONFIG_SLM_PGPS=n
-CONFIG_SLM_LOCATION=n
