--- applications/serial_lte_modem/src/gnss/Kconfig
@@ -1,34 +1,0 @@
-#
-# Copyright (c) 2021 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-
-config SLM_GNSS
-	bool "GNSS support in SLM"
-	default y
-
-if SLM_GNSS
-
-config SLM_AGPS
-	bool "Use nRF Cloud A-GPS"
-	depends on NRF_CLOUD_AGPS
-	depends on NRF_CLOUD_MQTT
-	help
-	  Use nRF Cloud A-GPS in GPS operation
-
-config SLM_PGPS
-	bool "Use nRF Cloud P-GPS"
-	depends on NRF_CLOUD_PGPS
-	depends on NRF_CLOUD_MQTT
-	help
-	  Use nRF Cloud P-GPS in GPS operation
-
-config SLM_LOCATION
-	bool "Use nRF Cloud Location"
-	depends on AT_MONITOR
-	depends on NRF_CLOUD_LOCATION
-	depends on NRF_CLOUD_MQTT
-	help
-	  Request nRF Cloud cellular or Wi-Fi location service
-
-endif # SLM_GNSS
