+++ applications/serial_lte_modem/src/slm_at_commands.c
@@ -29,6 +29,9 @@
 #include "slm_at_icmp.h"
 #include "slm_at_sms.h"
 #include "slm_at_fota.h"
+#if defined(CONFIG_SLM_NRF_CLOUD)
+#include "slm_at_nrfcloud.h"
+#endif
 #if defined(CONFIG_SLM_GNSS)
 #include "slm_at_gnss.h"
 #endif
@@ -415,16 +418,23 @@ int handle_at_sms(enum at_cmd_type cmd_type);
 /* FOTA commands */
 int handle_at_fota(enum at_cmd_type cmd_type);
 
-#if defined(CONFIG_SLM_GNSS)
-int handle_at_gps(enum at_cmd_type cmd_type);
+#if defined(CONFIG_SLM_NRF_CLOUD)
 int handle_at_nrf_cloud(enum at_cmd_type cmd_type);
-int handle_at_agps(enum at_cmd_type cmd_type);
-#if defined(CONFIG_SLM_PGPS)
-int handle_at_pgps(enum at_cmd_type cmd_type);
-#endif
-int handle_at_gps_delete(enum at_cmd_type cmd_type);
+#if defined(CONFIG_NRF_CLOUD_LOCATION)
 int handle_at_cellpos(enum at_cmd_type cmd_type);
 int handle_at_wifipos(enum at_cmd_type cmd_type);
+#endif
+#endif
+
+#if defined(CONFIG_SLM_GNSS)
+int handle_at_gps(enum at_cmd_type cmd_type);
+int handle_at_gps_delete(enum at_cmd_type cmd_type);
+#if defined(CONFIG_SLM_NRF_CLOUD) && defined(CONFIG_NRF_CLOUD_AGPS)
+int handle_at_agps(enum at_cmd_type cmd_type);
+#endif
+#if defined(CONFIG_SLM_NRF_CLOUD) && defined(CONFIG_NRF_CLOUD_PGPS)
+int handle_at_pgps(enum at_cmd_type cmd_type);
+#endif
 #endif
 
 #if defined(CONFIG_SLM_FTPC)
@@ -518,20 +528,23 @@ static struct slm_at_cmd {
 	/* FOTA commands */
 	{"AT#XFOTA", handle_at_fota},
 
+#if defined(CONFIG_SLM_NRF_CLOUD)
+	{"AT#XNRFCLOUD", handle_at_nrf_cloud},
+#if defined(CONFIG_NRF_CLOUD_LOCATION)
+	{"AT#XCELLPOS", handle_at_cellpos},
+	{"AT#XWIFIPOS", handle_at_wifipos},
+#endif
+#endif
+
 #if defined(CONFIG_SLM_GNSS)
 	/* GNSS commands */
 	{"AT#XGPS", handle_at_gps},
-	{"AT#XNRFCLOUD", handle_at_nrf_cloud},
-#if defined(CONFIG_SLM_AGPS)
+	{"AT#XGPSDEL", handle_at_gps_delete},
+#if defined(CONFIG_SLM_NRF_CLOUD) && defined(CONFIG_NRF_CLOUD_AGPS)
 	{"AT#XAGPS", handle_at_agps},
 #endif
-#if defined(CONFIG_SLM_PGPS)
+#if defined(CONFIG_SLM_NRF_CLOUD) && defined(CONFIG_NRF_CLOUD_PGPS)
 	{"AT#XPGPS", handle_at_pgps},
-#endif
-	{"AT#XGPSDEL", handle_at_gps_delete},
-#if defined(CONFIG_SLM_LOCATION)
-	{"AT#XCELLPOS", handle_at_cellpos},
-	{"AT#XWIFIPOS", handle_at_wifipos},
 #endif
 #endif
 
@@ -659,10 +672,17 @@ int slm_at_init(void)
 		LOG_ERR("FOTA could not be initialized: %d", err);
 		return -EFAULT;
 	}
+#if defined(CONFIG_SLM_NRF_CLOUD)
+	err = slm_at_nrfcloud_init();
+	if (err) {
+		LOG_ERR("nRF Cloud could not be initialized: %d", err);
+		return -EFAULT;
+	}
+#endif
 #if defined(CONFIG_SLM_GNSS)
 	err = slm_at_gnss_init();
 	if (err) {
-		LOG_ERR("GPS could not be initialized: %d", err);
+		LOG_ERR("GNSS could not be initialized: %d", err);
 		return -EFAULT;
 	}
 #endif
@@ -748,10 +768,16 @@ void slm_at_uninit(void)
 	if (err) {
 		LOG_WRN("FOTA could not be uninitialized: %d", err);
 	}
+#if defined(CONFIG_SLM_NRF_CLOUD)
+	err = slm_at_nrfcloud_uninit();
+	if (err) {
+		LOG_WRN("nRF Cloud could not be uninitialized: %d", err);
+	}
+#endif
 #if defined(CONFIG_SLM_GNSS)
 	err = slm_at_gnss_uninit();
 	if (err) {
-		LOG_WRN("GPS could not be uninitialized: %d", err);
+		LOG_WRN("GNSS could not be uninitialized: %d", err);
 	}
 #endif
 #if defined(CONFIG_SLM_FTPC)
